Fixed db recursive pub func to omit md
Changed db-get-publication-recursive-public to not include md.
This is required, because it uses the same ->jsexpr function as other
pub queries. We may want to include md in the API in the future. At
some point, we'll need to make separate API pub functions for that.
But for now, the main app takes precedences.